Login Issue Fixed

+++ REST-API/Server/controllers/feed.js
@@ -5,32 +5,31 @@ const { v4: uuidv4 } = require("uuid");
 const { validationResult } = require("express-validator");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
-exports.getPosts = (req, res, next) => {
+exports.getPosts = async (req, res, next) => {
     const currPage = req.query.page || 1;
     const perPage = 2;
-    let totalItems;
-    Post.find()
-      .countDocuments()
-      .then((count) => {
-        totalItems = count;
-        return Post.find()
-            .skip((currPage - 1) * perPage)
-            .limit(perPage);
-      })
-      .then((posts) => {
-        res.status(200).json({
-          message: "Fetched posts successfully.",
-          posts: posts,
-          totalItems: totalItems
-        });
-      })
-      .catch((err) => {
-        if (!err.statusCode) {
+    // let totalItems;
+    try {
+      const totalItems = await Post.find().countDocuments();
+      const posts = await Post.find()
+        .populate("creator")
+        .skip((currPage - 1) * perPage)
+        .limit(perPage);
+
+      res.status(200).json({
+        message: "Fetched posts successfully.",
+        posts: posts,
+        totalItems: totalItems
+      });
+    }
+    catch (err) {
+      if (!err.statusCode) {
           error.statusCode = 500;
         }
         next(err);
-      });
+    }
 };
 
 exports.createPost = (req, res, next) => {
@@ -50,24 +49,32 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\", "/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title, 
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'Anant Duhan',
-        }
+        creator: req.userId
     });
     post.save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
         res.status(201).json({
           message: "Post created successfully!",
-          post: result,
+          post: post,
+          creator: { _id: creator._id, name: creator.name }
         });
     })
     .catch(err => {
         if(!err.statusCode) {
-            error.statusCode = 500;
+            err.statusCode = 500;
         }
         next(err);
     });
@@ -121,6 +128,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
           }
+          if(post.creator.toString() !== req.userId) {
+            const error = new Error('Not Authorized');
+            error.statusCode = 422;
+            throw error;
+          }
           if(imageUrl !== post.imageUrl) {
               clearImage(post.imageUrl);
           }
@@ -149,13 +161,24 @@ exports.deletePost = (req, res, next) => {
               error.statusCode = 404;
               throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+              const error = new Error("Not Authorized");
+              error.statusCode = 422;
+              throw error;
+            }
             // check loggedIn User
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
-            res.status(200).json({ message: 'Deleted Post' });
+          return User.findById(req.userId);
+        })
+        .then(user => {
+          user.posts.pull(postId);
+          return user.save();
+        })
+        .then(result => {
+          res.status(200).json({ message: "Deleted Post" });
         })
         .catch((err) => {
             if (!err.statusCode) {
@@ -164,6 +187,54 @@ exports.deletePost = (req, res, next) => {
             next(err);
         });
 }
+
+exports.getStatus = async (req, res, next) => {
+  try {
+    const user = await User.findById(req.userId);
+
+    if(!user) {
+      const error = new Error('User not Exist');
+      error.statusCode = 404;
+      throw error;
+    }
+    res.status(200).json({
+      status: user.status
+    });
+  }
+  catch(error) {
+    if(!error.statusCode) {
+      error.statusCode = 500;
+    }
+    next(error);
+  }
+}
+
+exports.putStatus = async (req, res, next) => {
+  try {
+    const { status } = req.body;
+    const user = await User.findById(req.userId);
+
+    if (!user) {
+      const error = new Error("User not exist");
+      error.statusCode = 404;
+      throw error;
+    }
+
+    user.status = status;
+    const respose = await user.save();
+
+    res.status(200).json({
+      message: "Status updated!!",
+      status: respose.status,
+    });
+  } catch (error) {
+    if (!error.statusCode) {
+      error.statusCode = 500;
+    }
+
+    next(error);
+  }
+};
 
 const clearImage = filePath => {
     filePath = path.join(__dirname, '..', filePath);
